isolate test stub and implement Countable
